# Optimise floating point `is_finite` (2x) and `is_infinite` (1.6x).